+++ pkg/sidecar/services/authorization/client/client.go
@@ -70,9 +70,10 @@ type client struct {
 
 	cache *internalCache
 
-	policies         clientSet[*sidecarSvcAuthzTypes.Policy]
-	roles            clientSet[*sidecarSvcAuthzTypes.Role]
-	userRoleBindings clientSet[*sidecarSvcAuthzTypes.UserRoleBinding]
+	policies          clientSet[*sidecarSvcAuthzTypes.Policy]
+	roles             clientSet[*sidecarSvcAuthzTypes.Role]
+	userRoleBindings  clientSet[*sidecarSvcAuthzTypes.UserRoleBinding]
+	groupRoleBindings clientSet[*sidecarSvcAuthzTypes.UserRoleBinding]
 }
 
 func (c *client) Revision() uint64 {
@@ -80,7 +81,7 @@ func (c *client) Revision() uint64 {
 }
 
 func (c *client) Evaluate(ctx context.Context, req *pbAuthorizationV1.AuthorizationV1PermissionRequest) (*pbAuthorizationV1.AuthorizationV1PermissionResponse, error) {
-	groups := c.get().extractGroups(req.GetUser())
+	groups := c.get().extractGroups(req.GetUser(), req.GetRoles())
 
 	return groups.Evaluate(req)
 }
@@ -114,7 +115,7 @@ func (c *client) Ready(ctx context.Context) error {
 	c.lock.RLock()
 	defer c.lock.RUnlock()
 
-	return errors.Errors(c.policies.ready(), c.roles.ready(), c.userRoleBindings.ready(), util.BoolSwitch(c.cache == nil, errors.Errorf("nil internalCache"), nil))
+	return errors.Errors(c.policies.ready(), c.roles.ready(), c.userRoleBindings.ready(), c.groupRoleBindings.ready(), util.BoolSwitch(c.cache == nil, errors.Errorf("nil internalCache"), nil))
 }
 
 func (c *client) setRoles(items map[string]*sidecarSvcAuthzTypes.Role) {
@@ -131,7 +132,7 @@ func (c *client) setRoles(items map[string]*sidecarSvcAuthzTypes.Role) {
 	c.lock.Lock()
 	defer c.lock.Unlock()
 
-	cache := newCache(c.policies.items, cp, c.userRoleBindings.items)
+	cache := newCache(c.policies.items, cp, c.userRoleBindings.items, c.groupRoleBindings.items)
 
 	c.roles.set(cp)
 	c.cache = &cache
@@ -151,9 +152,29 @@ func (c *client) setUserRoleBindings(items map[string]*sidecarSvcAuthzTypes.User
 	c.lock.Lock()
 	defer c.lock.Unlock()
 
-	cache := newCache(c.policies.items, c.roles.items, cp)
+	cache := newCache(c.policies.items, c.roles.items, cp, c.groupRoleBindings.items)
 
 	c.userRoleBindings.set(cp)
+	c.cache = &cache
+}
+
+func (c *client) setGroupRoleBindings(items map[string]*sidecarSvcAuthzTypes.UserRoleBinding) {
+	c.setLock.Lock()
+	defer c.setLock.Unlock()
+
+	c.revision += 1
+
+	cp := make(map[string]*sidecarSvcAuthzTypes.UserRoleBinding, len(items))
+	for k, v := range items {
+		cp[k] = v
+	}
+
+	c.lock.Lock()
+	defer c.lock.Unlock()
+
+	cache := newCache(c.policies.items, c.roles.items, c.userRoleBindings.items, cp)
+
+	c.groupRoleBindings.set(cp)
 	c.cache = &cache
 }
 
@@ -171,7 +192,7 @@ func (c *client) setPolicies(items map[string]*sidecarSvcAuthzTypes.Policy) {
 	c.lock.Lock()
 	defer c.lock.Unlock()
 
-	cache := newCache(cp, c.roles.items, c.userRoleBindings.items)
+	cache := newCache(cp, c.roles.items, c.userRoleBindings.items, c.groupRoleBindings.items)
 
 	c.policies.set(cp)
 	c.cache = &cache
@@ -226,6 +247,10 @@ func (c *client) runE(ctx context.Context) error {
 
 	g.Go(func() error {
 		return c.runUserRoleBindingsE(gctx)
+	})
+
+	g.Go(func() error {
+		return c.runGroupRoleBindingsE(gctx)
 	})
 
 	return g.Wait()
@@ -437,6 +462,76 @@ func (c *client) runUserRoleBindingsE(ctx context.Context) error {
 			}
 
 			c.setUserRoleBindings(bindings)
+		}
+	}
+}
+
+func (c *client) runGroupRoleBindingsE(ctx context.Context) error {
+	bindings := map[string]*sidecarSvcAuthzTypes.UserRoleBinding{}
+
+	var index uint32
+
+	client, err := c.client.Get(ctx)
+	if err != nil {
+		return err
+	}
+
+	{
+		response, err := client.GetGroupRoleBinding(ctx, &pbSharedV1.Empty{})
+		if err != nil {
+			return err
+		}
+
+		for {
+			spec, err := response.Recv()
+			if err != nil {
+				if err == io.EOF {
+					break
+				}
+				return err
+			}
+
+			for _, item := range spec.GetItems() {
+				bindings[item.GetName()] = item.GetItem()
+				index = item.GetIndex()
+			}
+		}
+	}
+
+	c.setGroupRoleBindings(bindings)
+
+	logger.Trace("GroupRoleBindings init complete")
+
+	for {
+		changes, err := client.PoolGroupRoleBindingChanges(ctx, &sidecarSvcAuthzDefinition.AuthorizationPoolRequest{
+			Start:   index,
+			Timeout: durationpb.New(15 * time.Second),
+		})
+		if err != nil {
+			return err
+		}
+
+		for {
+			spec, err := changes.Recv()
+			if err != nil {
+				if err == io.EOF {
+					break
+				}
+				return err
+			}
+
+			logger.Int("items", len(spec.Items)).Trace("Received group role bindings update")
+
+			for _, item := range spec.GetItems() {
+				if item.GetItem() == nil || item.GetItem().Deleted() {
+					delete(bindings, item.GetName())
+				} else {
+					bindings[item.GetName()] = item.GetItem()
+				}
+				index = item.GetIndex()
+			}
+
+			c.setGroupRoleBindings(bindings)
 		}
 	}
 }